# perf: Antiforgery token validation & generation