--- a/src/Antiforgery/src/Internal/DefaultAntiforgeryTokenSerializer.cs
+++ b/src/Antiforgery/src/Internal/DefaultAntiforgeryTokenSerializer.cs
@@ -1,6 +1,8 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Buffers;
+using System.Diagnostics;
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.ObjectPool;
@@ -30,26 +32,30 @@ internal sealed class DefaultAntiforgeryTokenSerializer : IAntiforgeryTokenSeria
     {
         var serializationContext = _pool.Get();
 
+        byte[]? rented = null;
         Exception? innerException = null;
         try
         {
-            var count = serializedToken.Length;
-            var charsRequired = WebEncoders.GetArraySizeRequiredToDecode(count);
-            var chars = serializationContext.GetChars(charsRequired);
-            var tokenBytes = WebEncoders.Base64UrlDecode(
+            var tokenLength = serializedToken.Length;
+            var charsRequired = WebEncoders.GetArraySizeRequiredToDecode(tokenLength);
+
+            var chars = charsRequired < 128 ? stackalloc byte[128] : (rented = ArrayPool<byte>.Shared.Rent(charsRequired));
+            chars = chars[..charsRequired];
+
+            var decodedResult = WebEncoders.TryBase64UrlDecode(
                 serializedToken,
                 offset: 0,
-                buffer: chars,
-                bufferOffset: 0,
-                count: count);
+                count: tokenLength,
+                destination: chars,
+                out var bytesWritten);
+            Debug.Assert(decodedResult is true);
 
+            // when DataProtection obtains Span<>'ish APIs,
+            // we will just pass in chars directly. For now we need to allocate.
+            var tokenBytes = chars.Slice(0, bytesWritten).ToArray();
             var unprotectedBytes = _cryptoSystem.Unprotect(tokenBytes);
-            var stream = serializationContext.Stream;
-            stream.Write(unprotectedBytes, offset: 0, count: unprotectedBytes.Length);
-            stream.Position = 0L;
 
-            var reader = serializationContext.Reader;
-            var token = Deserialize(reader);
+            var token = Deserialize(unprotectedBytes);
             if (token != null)
             {
                 return token;
@@ -63,6 +69,11 @@ internal sealed class DefaultAntiforgeryTokenSerializer : IAntiforgeryTokenSeria
         finally
         {
             _pool.Return(serializationContext);
+
+            if (rented is not null)
+            {
+                ArrayPool<byte>.Shared.Return(rented);
+            }
         }
 
         // if we reached this point, something went wrong deserializing
@@ -81,39 +92,49 @@ internal sealed class DefaultAntiforgeryTokenSerializer : IAntiforgeryTokenSeria
      *   |    `- Username: UTF-8 string with 7-bit integer length prefix
      *   `- AdditionalData: UTF-8 string with 7-bit integer length prefix
      */
-    private static AntiforgeryToken? Deserialize(BinaryReader reader)
+    private static AntiforgeryToken? Deserialize(ReadOnlySpan<byte> unprotectedBytes)
     {
+        // pointer to current byte at unprotectedBytes
+        var pointer = 0;
+
         // we can only consume tokens of the same serialized version that we generate
-        var embeddedVersion = reader.ReadByte();
+        var embeddedVersion = unprotectedBytes[pointer++];
         if (embeddedVersion != TokenVersion)
         {
             return null;
         }
 
         var deserializedToken = new AntiforgeryToken();
-        var securityTokenBytes = reader.ReadBytes(AntiforgeryToken.SecurityTokenBitLength / 8);
-        deserializedToken.SecurityToken =
-            new BinaryBlob(AntiforgeryToken.SecurityTokenBitLength, securityTokenBytes);
-        deserializedToken.IsCookieToken = reader.ReadBoolean();
 
+        var securityTokenBytesLength = AntiforgeryToken.SecurityTokenBitLength / 8;
+        var securityTokenBytes = unprotectedBytes.Slice(pointer, securityTokenBytesLength).ToArray();
+        pointer += securityTokenBytesLength;
+        deserializedToken.SecurityToken = new BinaryBlob(AntiforgeryToken.SecurityTokenBitLength, securityTokenBytes);
+
+        deserializedToken.IsCookieToken = unprotectedBytes.BinaryReadBoolean(ref pointer);
         if (!deserializedToken.IsCookieToken)
         {
-            var isClaimsBased = reader.ReadBoolean();
+            var isClaimsBased = unprotectedBytes.BinaryReadBoolean(ref pointer);
             if (isClaimsBased)
             {
-                var claimUidBytes = reader.ReadBytes(AntiforgeryToken.ClaimUidBitLength / 8);
+                var claimUidBytesLength = AntiforgeryToken.ClaimUidBitLength / 8;
+                var claimUidBytes = unprotectedBytes.Slice(pointer, claimUidBytesLength).ToArray();
+                pointer += claimUidBytesLength;
                 deserializedToken.ClaimUid = new BinaryBlob(AntiforgeryToken.ClaimUidBitLength, claimUidBytes);
             }
             else
             {
-                deserializedToken.Username = reader.ReadString();
+                deserializedToken.Username = unprotectedBytes.Slice(pointer).BinaryReadString(out var usernameBytesRead);
+                pointer += usernameBytesRead;
             }
 
-            deserializedToken.AdditionalData = reader.ReadString();
+            deserializedToken.AdditionalData = unprotectedBytes.Slice(pointer).BinaryReadString(out var addDataBytesRead);
+            pointer += addDataBytesRead;
+
         }
 
         // if there's still unconsumed data in the stream, fail
-        if (reader.BaseStream.ReadByte() != -1)
+        if (pointer < unprotectedBytes.Length && unprotectedBytes[pointer] != 0x00)
         {
             return null;
         }