--- a/src/Antiforgery/src/Internal/IClaimUidExtractor.cs
+++ b/src/Antiforgery/src/Internal/IClaimUidExtractor.cs
@@ -11,6 +11,14 @@ namespace Microsoft.AspNetCore.Antiforgery;
 internal interface IClaimUidExtractor
 {
     /// <summary>
+    /// Extracts claims identifier and writes to <paramref name="destination"/>.
+    /// </summary>
+    /// <param name="claimsPrincipal">The <see cref="ClaimsPrincipal"/>.</param>
+    /// <param name="destination">destination for claimUid bytes</param>
+    /// <param name="bytesWritten">length of bytes written to <paramref name="destination"/></param>
+    void ExtractClaimUid(ClaimsPrincipal claimsPrincipal, Span<byte> destination, out int bytesWritten);
+
+    /// <summary>
     /// Extracts claims identifier.
     /// </summary>
     /// <param name="claimsPrincipal">The <see cref="ClaimsPrincipal"/>.</param>