--- a/src/Antiforgery/test/Microsoft.AspNetCore.Antiforgery.Test.csproj
+++ b/src/Antiforgery/test/Microsoft.AspNetCore.Antiforgery.Test.csproj
@@ -16,5 +16,4 @@
   <ItemGroup>
     <Compile Include="$(SharedSourceRoot)AntiforgeryMetadata.cs" />
   </ItemGroup>
-
 </Project>