… src/Antiforgery/test/DefaultAntiforgeryTest.cs
… src/Antiforgery/test/DefaultAntiforgeryTokenGeneratorTest.cs
… src/Antiforgery/test/DefaultClaimUidExtractorTest.cs
… src/Shared/test/Shared.Tests/WebEncodersTests.cs