**IN PROGRESS**

main stats for [benchmark](https://github.com/DeagleGross/Baraholka/blob/main/Benchmarks/AntiforgeryBenchmarks.cs)
| OS       |   Method           | Mean       | Error   | StdDev  | Gen0   | Gen1   | Gen2   | Allocated |
| -------- | ----------------- |-----------:|--------:|--------:|-------:|-------:|-------:|----------:|
| win      | Generate | 7,175.2 ns | 4.19 ns | 3.50 ns | 0.1297 |      - |      - |    4096 B |
| win      | ValidateWithUser |   768.0 ns | 2.52 ns | 2.35 ns | 0.0334 | 0.0010 | 0.0010 |         - |
| win      | Validate         |   144.4 ns | 0.21 ns | 0.18 ns | 0.0045 | 0.0005 | 0.0005 |     120 B  |


# {PR title}

<!-- Thank you for submitting a pull request to our repo. -->

<!-- If this is your first PR in the ASP.NET Core repo, please run through the checklist
below to ensure a smooth review and merge process for your PR. -->

- [ ] You've read the [Contributor Guide](https://github.com/dotnet/aspnetcore/blob/main/CONTRIBUTING.md) and [Code of Conduct](https://github.com/dotnet/aspnetcore/blob/main/CODE-OF-CONDUCT.md).
- [ ] You've included unit or integration tests for your change, where applicable.
- [ ] You've included inline docs for your change, where applicable.
- [ ] There's an open issue for the PR that you are making. If you'd like to propose a new feature or change, please open an issue to discuss the change or find an existing issue.

<!-- Once all that is done, you're ready to go. Open the PR with the content below. -->

Summary of the changes (Less than 80 chars)

## Description

{Detail}

Fixes #{bug number} (in this specific format)